split the canvas.js file into modules
