+++ src/application/drawing/UIElements.js
@@ -1,0 +1,14 @@
+export const canvas = document.querySelector("#main-canvas");
+export const drawingPreview = document.querySelector("#drawing-preview");
+
+export const lineWidthInput = document.querySelector("input#drawing-line-width");
+export const drawingColorInput = document.querySelector("input#drawing-line-color");
+
+export const resizer = document.querySelector("#resizer");
+
+export const drawingOptions = document.querySelector("#drawing-options");
+export const carac = document.querySelector("#caracteristics");
+
+export const fontsDatalist = document.querySelector("datalist");
+
+export const closeCanvasBtn = document.querySelector("#return");
